Struct graph_mock::store::MockStore

pub struct MockStore {
    ChainStore_expectations: MockStore_ChainStore,
    get_mock: Expectations,
    input_schema: Expectations,
    api_schema: Expectations,
    network_name: Expectations,
}

Fields
ChainStore_expectations: MockStore_ChainStoreget_mock: Expectationsinput_schema: Expectationsapi_schema: Expectationsnetwork_name: Expectations
Implementations

impl MockStore

pub fn get_mock(
&self,
key: EntityKey
) -> Result<Option<Entity>, QueryExecutionError>

pub fn input_schema(
&self,
subgraph_id: &DeploymentHash
) -> Result<Arc<Schema>, StoreError>

pub fn api_schema(
&self,
subgraph_id: &DeploymentHash
) -> Result<Arc<ApiSchema>, StoreError>

pub fn network_name(
&self,
subgraph_id: &DeploymentHash
) -> Result<Option<String>, StoreError>

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_get_mock(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_mock method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_input_schema(&mut self) -> &mut Expectation

Create an Expectation for mocking the input_schema method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_api_schema(&mut self) -> &mut Expectation

Create an Expectation for mocking the api_schema method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_network_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the network_name method

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockStore

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_genesis_block_ptr(&mut self) -> &mut Expectation

Create an Expectation for mocking the genesis_block_ptr method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_upsert_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the upsert_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_upsert_light_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the upsert_light_blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_attempt_chain_head_update(&mut self) -> &mut Expectation

Create an Expectation for mocking the attempt_chain_head_update method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_chain_head_ptr(&mut self) -> &mut Expectation

Create an Expectation for mocking the chain_head_ptr method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_ancestor_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the ancestor_block method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_cleanup_cached_blocks(&mut self) -> &mut Expectation

Create an Expectation for mocking the cleanup_cached_blocks method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_hashes_by_block_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_hashes_by_block_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_confirm_block_hash(&mut self) -> &mut Expectation

Create an Expectation for mocking the confirm_block_hash method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_block_number(&mut self) -> &mut Expectation

Create an Expectation for mocking the block_number method

#[must_use = "Must set return value when not using the \"nightly\" feature"]pub fn expect_transaction_receipts_in_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the transaction_receipts_in_block method